Blockchain/requirements.txt
Flask
pymongo
python-dotenv
requests
PyJWT
cryptography
